noto: add ability for notary to use a "guard" contract
For programmable approval logic, a notary  may choose to deploy a private
"guard" contract using Pente. In this mode, all Noto transactions will be
forwarded via this contract in order to be submitted to the base ledger.

Note that this is not yet tested E2E with Pente, but is partially
simulated using a base ledger "guard" contract.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>